backend/requirements.txt
Flask == 2.2.3
Flask-Cors == 3.0.10
Werkzeug==2.2.2
python-dotenv == 1.0.0
PyPDF2 == 3.0.1
langchain == 0.1.0
openai == 1.7.1
tiktoken == 0.5.2
faiss-cpu == 1.7.4
